Refactor :: 컴포넌트의 디렉토리 구조 재구성 #133
Joined 관련 파일들을 chat.member 이동하고, room 관련 파일들을 chat.room으로 이동하는 등 파일의 구조를 전반적으로 조정하였습니다.
또한, 필요한 부분에서 패키지 임포트 경로를 적절히 수정하였습니다.
