+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/chat/member/dto/request/AddJoinedRequest.kt
@@ -1,4 +1,4 @@
-package com.seugi.api.domain.chat.presentation.joined.dto.request
+package com.seugi.api.domain.chat.presentation.chat.member.dto.request
 
 data class AddJoinedRequest(
     val chatRoomId: Long? = null,
