+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/chat/member/dto/request/TossMasterRequest.kt
@@ -1,4 +1,4 @@
-package com.seugi.api.domain.chat.presentation.joined.dto.request
+package com.seugi.api.domain.chat.presentation.chat.member.dto.request
 
 data class TossMasterRequest(
     val roomId: Long,
