+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/chat/room/controller/GroupChatController.kt
@@ -1,24 +1,23 @@
-package com.seugi.api.domain.chat.presentation.room.controller
+package com.seugi.api.domain.chat.presentation.chat.room.controller
 
-import com.seugi.api.domain.chat.application.service.room.ChatRoomService
+import com.seugi.api.domain.chat.application.service.chat.room.ChatRoomService
 import com.seugi.api.domain.chat.domain.enums.type.RoomType
-import com.seugi.api.domain.chat.presentation.room.dto.request.CreateRoomRequest
-import com.seugi.api.domain.chat.presentation.room.dto.request.SearchRoomRequest
-import com.seugi.api.domain.chat.presentation.room.dto.response.RoomResponse
+import com.seugi.api.domain.chat.presentation.chat.room.dto.request.CreateRoomRequest
+import com.seugi.api.domain.chat.presentation.chat.room.dto.request.SearchRoomRequest
+import com.seugi.api.domain.chat.presentation.chat.room.dto.response.RoomResponse
 import com.seugi.api.global.common.annotation.GetAuthenticatedId
 import com.seugi.api.global.response.BaseResponse
 import org.springframework.web.bind.annotation.*
 
-
 /**
- * 개인
- * 방 생성, 방 찾기, 방에대한 채팅내역 불러오기
+ * 그룹
+ * 방 생성 or 삭제 or 수정(이름, 사진), 방 찾기, 방에대한 채팅내역 불러오기
  */
 
 @RestController
-@RequestMapping("/chat/personal")
-class PersonalChatController(
-    private val chatRoomService: ChatRoomService
+@RequestMapping("/chat/group")
+class GroupChatController(
+    private val chatRoomService: ChatRoomService,
 ) {
 
     @PostMapping("/create")
@@ -26,7 +25,7 @@ class PersonalChatController(
         @GetAuthenticatedId id: Long,
         @RequestBody createRoomRequest: CreateRoomRequest
     ): BaseResponse<String> {
-        return chatRoomService.createChatRoom(createRoomRequest, id, RoomType.PERSONAL)
+        return chatRoomService.createChatRoom(createRoomRequest, id, RoomType.GROUP)
     }
 
     @GetMapping("/search")
@@ -35,10 +34,9 @@ class PersonalChatController(
         @RequestParam("workspace", defaultValue = "") workspaceId: String,
         @RequestParam("word", defaultValue = "") word: String
     ): BaseResponse<List<RoomResponse>> {
-
         return chatRoomService.searchRoomNameIn(
             SearchRoomRequest(workspaceId = workspaceId, word = word),
-            RoomType.PERSONAL,
+            RoomType.GROUP,
             userId
         )
     }
@@ -54,12 +52,22 @@ class PersonalChatController(
         )
     }
 
-    @GetMapping("/search/{workspaceId}")
+    @GetMapping("/search/{workspaceID}")
     fun getRooms(
         @GetAuthenticatedId userid: Long,
-        @PathVariable workspaceId: String
+        @PathVariable workspaceID: String,
     ): BaseResponse<List<RoomResponse>> {
-        return chatRoomService.getRooms(workspaceId, userid, RoomType.PERSONAL)
+        return chatRoomService.getRooms(workspaceID, userid, RoomType.GROUP)
+    }
+
+
+    //나가기
+    @PatchMapping("/left/{roomId}")
+    fun leftRoom(
+        @GetAuthenticatedId userId: Long,
+        @PathVariable roomId: String,
+    ): BaseResponse<Unit> {
+        return chatRoomService.leftRoom(userId, roomId)
     }
 
 }
