+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/chat/room/controller/PersonalChatController.kt
@@ -1,23 +1,24 @@
-package com.seugi.api.domain.chat.presentation.room.controller
+package com.seugi.api.domain.chat.presentation.chat.room.controller
 
-import com.seugi.api.domain.chat.application.service.room.ChatRoomService
+import com.seugi.api.domain.chat.application.service.chat.room.ChatRoomService
 import com.seugi.api.domain.chat.domain.enums.type.RoomType
-import com.seugi.api.domain.chat.presentation.room.dto.request.CreateRoomRequest
-import com.seugi.api.domain.chat.presentation.room.dto.request.SearchRoomRequest
-import com.seugi.api.domain.chat.presentation.room.dto.response.RoomResponse
+import com.seugi.api.domain.chat.presentation.chat.room.dto.request.CreateRoomRequest
+import com.seugi.api.domain.chat.presentation.chat.room.dto.request.SearchRoomRequest
+import com.seugi.api.domain.chat.presentation.chat.room.dto.response.RoomResponse
 import com.seugi.api.global.common.annotation.GetAuthenticatedId
 import com.seugi.api.global.response.BaseResponse
 import org.springframework.web.bind.annotation.*
 
+
 /**
- * 그룹
- * 방 생성 or 삭제 or 수정(이름, 사진), 방 찾기, 방에대한 채팅내역 불러오기
+ * 개인
+ * 방 생성, 방 찾기, 방에대한 채팅내역 불러오기
  */
 
 @RestController
-@RequestMapping("/chat/group")
-class GroupChatController(
-    private val chatRoomService: ChatRoomService,
+@RequestMapping("/chat/personal")
+class PersonalChatController(
+    private val chatRoomService: ChatRoomService
 ) {
 
     @PostMapping("/create")
@@ -25,7 +26,7 @@ class GroupChatController(
         @GetAuthenticatedId id: Long,
         @RequestBody createRoomRequest: CreateRoomRequest
     ): BaseResponse<String> {
-        return chatRoomService.createChatRoom(createRoomRequest, id, RoomType.GROUP)
+        return chatRoomService.createChatRoom(createRoomRequest, id, RoomType.PERSONAL)
     }
 
     @GetMapping("/search")
@@ -34,9 +35,10 @@ class GroupChatController(
         @RequestParam("workspace", defaultValue = "") workspaceId: String,
         @RequestParam("word", defaultValue = "") word: String
     ): BaseResponse<List<RoomResponse>> {
+
         return chatRoomService.searchRoomNameIn(
             SearchRoomRequest(workspaceId = workspaceId, word = word),
-            RoomType.GROUP,
+            RoomType.PERSONAL,
             userId
         )
     }
@@ -52,22 +54,12 @@ class GroupChatController(
         )
     }
 
-    @GetMapping("/search/{workspaceID}")
+    @GetMapping("/search/{workspaceId}")
     fun getRooms(
         @GetAuthenticatedId userid: Long,
-        @PathVariable workspaceID: String,
+        @PathVariable workspaceId: String
     ): BaseResponse<List<RoomResponse>> {
-        return chatRoomService.getRooms(workspaceID, userid, RoomType.GROUP)
-    }
-
-
-    //나가기
-    @PatchMapping("/left/{roomId}")
-    fun leftRoom(
-        @GetAuthenticatedId userId: Long,
-        @PathVariable roomId: String,
-    ): BaseResponse<Unit> {
-        return chatRoomService.leftRoom(userId, roomId)
+        return chatRoomService.getRooms(workspaceId, userid, RoomType.PERSONAL)
     }
 
 }
