+++ src/main/kotlin/com/seugi/api/domain/chat/presentation/chat/room/dto/response/RoomResponse.kt
@@ -1,4 +1,4 @@
-package com.seugi.api.domain.chat.presentation.room.dto.response
+package com.seugi.api.domain.chat.presentation.chat.room.dto.response
 
 import com.seugi.api.domain.chat.domain.enums.status.ChatStatusEnum
 import com.seugi.api.domain.chat.domain.enums.type.RoomType
